+++ health-services/libraries/health-services-common/src/main/java/org/egov/common/data/query/builder/SelectQueryBuilder.java
@@ -1,5 +1,6 @@
 package org.egov.common.data.query.builder;
 
+import org.apache.commons.lang3.StringUtils;
 import org.egov.common.data.query.exception.QueryBuilderException;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
@@ -20,9 +21,19 @@ public class SelectQueryBuilder implements GenericQueryBuilder {
 
     @Override
     public String build(Object object) throws QueryBuilderException {
+        String tableName = null;
+        try {
+            tableName = GenericQueryBuilder.getTableName(object.getClass());
+        } catch (Exception exception) {
+            throw new QueryBuilderException(exception.getMessage());
+        }
+
+        return build(object, tableName);
+    }
+
+    public String build(Object object, String tableName) throws QueryBuilderException {
         StringBuilder queryStringBuilder = null;
         try {
-            String tableName = GenericQueryBuilder.getTableName(object.getClass());
             List<String> whereClauses = GenericQueryBuilder.getFieldsWithCondition(object, QueryFieldChecker.isNotNull, paramsMap);
             queryStringBuilder = GenericQueryBuilder.generateQuery(GenericQueryBuilder.selectQueryTemplate(tableName), whereClauses);
         } catch (Exception exception) {