+++ health-services/facility/CHANGELOG.md
@@ -1,5 +1,10 @@
 All notable changes to this module will be documented in this file.
 
+## 1.1.2 - 2024-05-29
+- Integrated Core 2.9LTS
+- Upgraded to health models 1.0.20 and health common 1.0.16
+- Integrated Boundary v2 functionality
+
 ## 1.0.0
 
 - Base version